Service classes for the Spring entities made in #11
