+++ .gitignore
@@ -1,3 +1,5 @@
 .idea/
 File-Integrity-Scanner/.idea/
 File-Integrity-Scanner/target/
+
+.vscode
