+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/msr/domain/model/entities/BaseEntity.java
@@ -1,0 +1,7 @@
+package org.pwss.file_integrity_scanner.msr.domain.model.entities;
+
+/**
+ * Limit the amount of generic classes the Base Service would accept
+ */
+public abstract class BaseEntity {
+}
