+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/msr/service/BaseService.java
@@ -1,0 +1,22 @@
+package org.pwss.file_integrity_scanner.msr.service;
+
+
+
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public abstract class BaseService<Repository extends JpaRepository> {
+
+    final Repository repository;
+
+  
+    public BaseService(@Autowired Repository repository) {
+
+        this.repository = repository;
+
+    }
+
+
+
+}
