+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/msr/service/checksum/ChecksumService.java
@@ -1,0 +1,6 @@
+package org.pwss.file_integrity_scanner.msr.service.checksum;
+
+
+public interface ChecksumService {
+
+}
