+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/msr/service/file/FileService.java
@@ -1,0 +1,5 @@
+package org.pwss.file_integrity_scanner.msr.service.file;
+
+public interface FileService {
+    
+}
