+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/msr/service/monitored_directory/MonitoredDirectoryService.java
@@ -1,0 +1,5 @@
+package org.pwss.file_integrity_scanner.msr.service.monitored_directory;
+
+public interface MonitoredDirectoryService {
+    
+}
