+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/msr/service/scan/ScanServiceImpl.java
@@ -1,0 +1,13 @@
+package org.pwss.file_integrity_scanner.msr.service.scan;
+import org.pwss.file_integrity_scanner.msr.repository.ScanRepository;
+import org.pwss.file_integrity_scanner.msr.service.BaseService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ScanServiceImpl extends BaseService<ScanRepository> implements ScanService {
+
+    public ScanServiceImpl(ScanRepository repository){
+    super(repository);
+}
+    
+}
