+++ File-Integrity-Scanner/src/main/java/org/pwss/file_integrity_scanner/msr/service/scan_details/ScanDetailsServiceImpl.java
@@ -1,0 +1,14 @@
+package org.pwss.file_integrity_scanner.msr.service.scan_details;
+
+import org.pwss.file_integrity_scanner.msr.repository.ChecksumRepository;
+import org.pwss.file_integrity_scanner.msr.repository.ScanDetailsRepository;
+import org.pwss.file_integrity_scanner.msr.service.BaseService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ScanDetailsServiceImpl extends BaseService<ScanDetailsRepository> implements ScanDetailsService {
+
+    public ScanDetailsServiceImpl(ScanDetailsRepository repository) {
+        super(repository);
+    }
+}
