--- File-Integrity-Scanner/src/main/java/org/pwss/msr/domain/model/entities/BaseEntity.java
@@ -1,7 +1,0 @@
-package org.pwss.msr.domain.model.entities;
-
-/**
- * Limit the amount of generic classes the Base Service would accept
- */
-public abstract class BaseEntity {
-}
