--- File-Integrity-Scanner/src/main/java/org/pwss/msr/domain/model/entities/scan/Scan.java
@@ -1,71 +1,0 @@
-package org.pwss.msr.domain.model.entities.scan;
-
-import jakarta.persistence.*;
-import org.pwss.msr.domain.model.entities.BaseEntity;
-import org.pwss.msr.domain.model.entities.monitored_directory.MonitoredDirectory;
-
-import java.time.OffsetDateTime;
-
-@Entity
-@Table(name = "scans")
-public class Scan extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Column(name = "scan_time", nullable = false)
-    private OffsetDateTime scanTime;
-
-    @Column(nullable = false)
-    private String status;
-
-    @Column
-    private String notes;
-
-    @ManyToOne(optional = false)
-    @JoinColumn(name = "monitored_directory_id", nullable = false)
-    private MonitoredDirectory monitoredDirectory;
-
-    // Getters and setters
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public OffsetDateTime getScanTime() {
-        return scanTime;
-    }
-
-    public void setScanTime(OffsetDateTime scanTime) {
-        this.scanTime = scanTime;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public void setStatus(String status) {
-        this.status = status;
-    }
-
-    public String getNotes() {
-        return notes;
-    }
-
-    public void setNotes(String notes) {
-        this.notes = notes;
-    }
-
-    public MonitoredDirectory getMonitoredDirectory() {
-        return monitoredDirectory;
-    }
-
-    public void setMonitoredDirectory(MonitoredDirectory monitoredDirectory) {
-        this.monitoredDirectory = monitoredDirectory;
-    }
-}
