--- File-Integrity-Scanner/src/main/java/org/pwss/msr/repository/ChecksumRepository.java
@@ -1,9 +1,0 @@
-package org.pwss.msr.repository;
-
-import org.pwss.msr.domain.model.entities.checksum.Checksum;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChecksumRepository extends JpaRepository<Checksum, Long> {
-}
