--- File-Integrity-Scanner/src/main/java/org/pwss/msr/repository/MonitoredDirectoryRepository.java
@@ -1,7 +1,0 @@
-package org.pwss.msr.repository;
-
-import org.pwss.msr.domain.model.entities.monitored_directory.MonitoredDirectory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MonitoredDirectoryRepository extends JpaRepository<MonitoredDirectory, Integer> {
-}
