+++ File-Integrity-Scanner/src/test/java/org/pwss/file_integrity_scanner/msr/service/checksum/ChecksumServiceTest.java
@@ -1,0 +1,22 @@
+package org.pwss.file_integrity_scanner.msr.service.checksum;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import org.junit.jupiter.api.Test;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class ChecksumServiceTest {
+
+    @Autowired 
+    ChecksumService checksumService;
+
+
+    @Test
+    public void testUserServiceInitialization() {
+    assertNotNull(checksumService, "ChecksumService should not be null");
+    }
+
+}
